+++ app/controllers/tasks/usage/user_activity_controller.rb
@@ -4,7 +4,7 @@ class Tasks::Usage::UserActivityController < ApplicationController
 
   include TaskControllerConfiguration
   def index
-    @project_members =  User.all # Project.find(@sessions_current_project_id).users
+    @project_members = User.all # Project.find(@sessions_current_project_id).users
   end
 
   def report